add some checking, start backup script

+++ clear.sh
@@ -2,10 +2,15 @@
 
 # clear.sh
 
-cd ../${1}
+if [ ${1} ];
+then
+    cd ../${1}
 
-echo 'Removing *.pyc ...'
-find -iname '*.pyc' -print0 | xargs -0 rm -rf
-echo 'Removing *.orig ...'
-find -iname '*.orig' -print0 | xargs -0 rm -rf
-
+    echo 'Removing *.pyc ...'
+    find -iname '*.pyc' -print0 | xargs -0 rm -rf
+    echo 'Removing *.orig ...'
+    find -iname '*.orig' -print0 | xargs -0 rm -rf
+else
+    echo Usage ${0} %%project_name%%
+    exit;
+fi
